# Conflicts:
#	drivers/can/can_mcan.c
#	drivers/can/can_mcp251xfd.c
#	drivers/can/can_mcp251xfd.h
#	drivers/can/can_sam0.c
#	drivers/modem/modem_cellular.c
#	drivers/wifi/esp_at/esp.c
#	drivers/wifi/esp_at/esp.h
#	include/zephyr/drivers/can.h
#	include/zephyr/drivers/can/can_mcan.h
#	include/zephyr/fs/fs.h
#	include/zephyr/net/net_if.h
#	include/zephyr/rtio/rtio.h
#	include/zephyr/smf.h
#	kernel/sched.c
#	lib/smf/smf.c
#	soc/arm/atmel_sam0/common/Kconfig.samc2x
#	soc/atmel/sam/samv71/soc.c
#	soc/atmel/sam/samv71/soc.h
#	soc/atmel/sam/samv71/soc_config.c
#	soc/atmel/sam0/common/dac_fixup_sam0.h
#	soc/atmel/sam0/samc21/soc.h
#	subsys/fs/nvs/nvs.c
#	subsys/modem/modem_cmux.c
#	subsys/modem/modem_ppp.c
#	subsys/net/ip/net_mgmt.c
#	subsys/net/lib/config/init_clock_sntp.c
#	subsys/shell/shell.c
#	west.yml

+++ subsys/net/ip/net_pkt.c
@@ -791,7 +791,7 @@ void net_pkt_compact(struct net_pkt *pkt)
 				frag->frags->data + copy_len,
 				frag->frags->len - copy_len);
 
-			frag->frags->len -= copy_len;
+			frag->frags->len -= (uint16_t)copy_len;
 
 			/* Is there any more space in this fragment */
 			if (net_buf_tailroom(frag)) {
@@ -886,7 +886,7 @@ static struct net_buf *pkt_alloc_buffer(struct net_buf_pool *pool,
 
 		current = new;
 		if (current->size > size) {
-			current->size = size;
+			current->size = (uint16_t)size;
 		}
 
 		size -= current->size;
@@ -1107,7 +1107,7 @@ int net_pkt_remove_tail(struct net_pkt *pkt, size_t length)
 
 	while (buf) {
 		if (buf->len >= remaining_len) {
-			buf->len = remaining_len;
+			buf->len = (uint16_t)remaining_len;
 
 			if (buf->frags) {
 				net_pkt_frag_unref(buf->frags);
@@ -1473,7 +1473,7 @@ pkt_alloc_with_buffer(struct k_mem_slab *slab,
 		return NULL;
 	}
 
-	net_pkt_set_family(pkt, family);
+	net_pkt_set_family(pkt, (uint8_t)family);
 
 	timeout = sys_timepoint_timeout(end);
 #if NET_LOG_LEVEL >= LOG_LEVEL_DBG
@@ -2034,7 +2034,7 @@ int net_pkt_update_length(struct net_pkt *pkt, size_t length)
 		if (buf->len < length) {
 			length -= buf->len;
 		} else {
-			buf->len = length;
+			buf->len = (uint16_t)length;
 			length = 0;
 		}
 	}
@@ -2065,7 +2065,7 @@ int net_pkt_pull(struct net_pkt *pkt, size_t length)
 			rem = length;
 		}
 
-		c_op->buf->len -= rem;
+		c_op->buf->len -= (uint16_t)rem;
 		left -= rem;
 		if (left) {
 			memmove(c_op->pos, c_op->pos+rem, left);
